legacy/root_debris/requirements.txt
Django>=4.2,<5.0
django-jazzmin>=2.6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
deep-translator>=1.11.4
requests>=2.31.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# AI/ML dependencies
librosa>=0.10.0
numpy==1.22.0
soundfile>=0.12.0
torch>=2.0.0
transformers>=4.33.0
TTS>=0.22.0
pydub>=0.25.1
